requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pydantic>=2.0.0
streamlit>=1.28.0

